Remove individual ExplorerDataSource implementations
The individual implementations of ExplorerDataSource are the same
implementation which is to return Unimplemented.  This could just be
the default implementation of ExplorerDataSource.

This commit adds a default implementation for ExplorerDataSource and
removes the unnecessary individual implementations.
